+++ lib/Utils/src/Utils.h
@@ -3,5 +3,5 @@
 void padRight(char padChar, char* str, int max);
 void padLeft(char padChar, char* str, int max);
 
-void metricfy(long value, char * buffer);
+void metricfy(float value, char * buffer);
 void timify(long time, char * buffer);
